refactor: moving ttc_control to an exclusive file and restoring controller.c to branch develop version
